record all measurements into one dep inj, remove unused code
